+++ src/azure-firewall/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+2.2.0
+++++++
+* `az network firewall policy create` : Updated parameter `--identity` to support multiple MSIs
+* `az network firewall policy update` : Updated parameter `--identity` to support multiple MSIs
+
 2.1.1
 ++++++
 * Update AzureFirewall model to support extended location.